+++ app/(dashboard)/dashboard/billing/page.tsx
@@ -1,9 +1,8 @@
 import { redirect } from "next/navigation"
-import { User } from "@prisma/client"
 
+import { env } from "@/env.mjs"
 import { PageAnalytics } from "@/lib/analytics-client"
 import { authOptions } from "@/lib/auth"
-import { db } from "@/lib/db"
 import { getCurrentUser } from "@/lib/session"
 import { stripe } from "@/lib/stripe"
 import { getUserSubscriptionPlan } from "@/lib/subscription"
@@ -11,24 +10,12 @@ import { BillingForm } from "@/components/billing-form"
 import { DashboardHeader } from "@/components/header"
 import PostHogAnalytics from "@/components/posthog"
 import { DashboardShell } from "@/components/shell"
+import { getTokensByUserId } from "@/app/api/users/utils"
 
 export const metadata = {
+  metadataBase: new URL(`${env.NEXT_PUBLIC_APP_URL}/dashboard/billing`),
   title: "Billing",
   description: "Manage billing and your subscription plan.",
-}
-
-export async function getTokensByUserId(id: User["id"]) {
-  const user = await db.user.findFirst({
-    where: {
-      id: id,
-    },
-  })
-
-  if (!user) {
-    redirect(authOptions?.pages?.signIn || "/login")
-  }
-
-  return Number(user.tokens)
 }
 
 export default async function BillingPage() {