+++ app/api/projects/[projectId]/keywords/[keywordId]/translations/route.ts
@@ -7,8 +7,8 @@ import { handleCatchApi } from "@/lib/exceptions"
 import i18n from "@/lib/i18n"
 import { ErrorResponse } from "@/lib/response"
 
-import { verifyCurrentUserHasAccessToProject } from "../../../route"
-import { routeContextSchemaProjectKeyword } from "../route"
+import { verifyCurrentUserHasAccessToProject } from "../../../utils"
+import { routeContextSchemaProjectKeyword } from "../utils"
 
 const translationCreateSchema = z.object({
   translations: z.array(