+++ app/api/projects/[projectId]/keywords/[keywordId]/utils.ts
@@ -1,0 +1,8 @@
+import { z } from "zod"
+
+export const routeContextSchemaProjectKeyword = z.object({
+  params: z.object({
+    projectId: z.string(),
+    keywordId: z.string(),
+  }),
+})